Moved scraper.py to `scraping` dir
